Fix Files System
resolves #5
