+++ Formula 1/Views/UIKit/Cells/ConstructorCell.swift
